refactor: enforce memory safety, 64-bit file format
Memory & Pipeline Safety:
- RAII File Management: Introduced ScopedCryFile (with deleted copy constructors) to guarantee FILE* handle closure during early returns or exceptions, preventing OS-level file locks.
- OOM Protection: Wrapped massive memory allocations (std::vector buffers, QImage) in try-catch blocks to gracefully handle std::bad_alloc, propagating readable error states instead of crashing.
- 64-bit Overflow Prevention: Migrated flat array indices, byte totals, and grid calculations from uint32 to size_t to prevent integer overflow on maps exceeding 32-bit limits.

Data-Oriented Design (DOD):
- Eliminated heap allocations by replacing std::string in SVegetationInstanceData with a 1-byte EVegetationLayers enum. Classification is now resolved upstream during editor queries.

ABI & LFS Scaling:
- Upgraded STileEntry::byteSize to uint64 to support chunk sizes exceeding 4GB (struct size updated to 20 bytes).
- Integrated CLFSFacade::GetMaxFileSize() pre-calculation to fail-fast before executing expensive baking operations if physical limits are breached.
- Aligned tileSize configurations to uint32 uniformly across settings and UI layers.

UI & Stability:
- Added QMessageBox critical dialogs to isolate business logic errors from UI representation.
- Extended MapPreview zoom limit to 64.0x for precise inspection of sparse pixel data.
- Fixed division-by-zero vulnerabilities in UI SpinBoxes by enforcing safe terrain bounds fallbacks.
- Hardened ConvertUtils with NaN/Inf validation for floating-point values.

+++ Settings/BakerSettings.h
@@ -6,7 +6,7 @@ namespace JDKLevelMaps::Settings
 	struct SBakerSettings
 	{
 		float cellSize = 1.0f;
-		uint16 tileSize = 64;
+		uint32 tileSize = 64;
 		SVegetationBakerSettings vegSettings;
 	};
 }
